Update README.md
Using the provided dataset, I am tasked with training a model to predict the type of cyberattack based on the input. I will follow the steps of a machine learning pipeline, to build and deploy a small web application. This application will feature an interface where users can input data, and it will return predictions for the type of cyberattack.